Parse position, end and length as i32 numbers

+++ src/models/cytoband.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone)]
 pub struct Cytoband {
-    pub start: u64,
-    pub end: u64,
+    pub start: i32,
+    pub end: i32,
     pub name: String,
 }
